# docs/swagger.yaml
basePath: /
definitions:
  basis_internal_dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
...
  basis_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  basis_internal_dto.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@basis.local
    name: API Support
  description: REST API сервиса управления задачами и командной аналитики.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Task Management System API
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные пользователя и возвращает JWT-токен.
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/basis_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/basis_internal_dto.AuthResponse'
        "400":
          description: Невалидное тело запроса или ошибки валидации
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе с хешированием пароля.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/basis_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Невалидное тело запроса или ошибки валидации
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Регистрация нового пользователя
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <JWT_TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"